WIP: TO CHECK!!! quic-be: Start asap the mux for 0-RTT
The mux is created and woken up as soon as possible. This means as soon as
the TX early data encryption level keys are installed. These keys are
installed on the first SSL_do_handshake() call. This call is done as soon
as the connection is created.

This way the mux is able to send 0-RTT packets alongside Initial packets.