WIP/MINOR: mux-h2: Save the known length of the payload
The former <body_len> field in the H2 stream is now called <curr_len>. It is
only used to check the payload length against the announced
content-length. <body_len> field is reused to save the known length of the
payload. It will be used to be able to properly define the known input
payload length value of the sedesc.

  NOTE: The name is confusing. A better name must be found.