BUG/MINOR: compression: Test payload size only if a content length was specified
When a minimum size is defined to performe the comression, the message
payload size is tested. To do so, information from the HTX message a used to
determine the message length. However it is performed regardless the payload
length is fully known or not. Concretely, the test must on be performed when
a content-length value was speficied. Otherwise, we are unable to really
determine the payload length.

Because of this bug, a large chunked message may be uncompressed because the
chunks already received are too small. But this does not mean the whole
message is small.

This patch must be backported to 3.2.